// frontend/src/metabase/query_builder/components/Canvas/types.ts
export type ProgrammingLanguageOptions =
  | "typescript"
  | "javascript"
  | "cpp"
  | "java"
  | "php"
  | "python"
  | "html"
  | "sql"
  | "other";
// ...
export interface Reflections {
  /**
   * Style rules to follow for generating content.
   */
  styleRules: string[];
  /**
   * Key content to remember about the user when generating content.
   */
  content: string[];
}
// ...
export interface ArtifactV3 {
  currentIndex: number;
  contents: (ArtifactMarkdownV3 | ArtifactCodeV3 | ArtifactQueryV3 | ArtifactInsightsV3)[];
}

export interface ArtifactMarkdownV3 {
  index: number;
  type: "text";
  title: string;
  description?: string;
  fullMarkdown: string;
}

export interface ArtifactCodeV3 {
  index: number;
  type: "code";
  title: string;
  description?: string;
  language: ProgrammingLanguageOptions;
  code: string;
}

export interface ArtifactQueryV3 {
  index: number;
  type: "card";
  title: string;
  cardId: any;
  queryResult?: any;
  rawResponse?: any;
}

export interface ArtifactImageV3 {
  index: number;
  type: "image";
  title: string;
  image: any;
}

export interface ArtifactInsightsV3 {
  index: number;
  type: "insight";
  title: string;
  insight: any;
}

export interface CodeHighlight {
  startCharIndex: number;
  endCharIndex: number;
}

export interface TextHighlight {
  fullMarkdown: string;
  markdownBlock: string;
  selectedText: string;
}